add lnurlerrorhandler only on lnurl routes
